add service bridge and rename
