+++ config/client.yaml
@@ -2,28 +2,28 @@
   ros__parameters:
     cloud_ip: "127.0.0.1"
     manage_port: 25565
-    # sub_list:
-    # - scan
-    # - tf_static
-    # - odom
-    # - imu
-    # - initialpose
-    # - move_simple
-    # - clock
-    # - bt_elev
-    # - bt_elev_cancel
-    # - color_compressed
-    # - move_simple
-    # - goal_pose
-    # tf_list:
-    # - odom_basefootprint
-    # srv_server_list:
-    # - load_map
+    sub_list:
+    - scan
+    - tf_static
+    - odom
+    - imu
+    - initialpose
+    - move_simple
+    - clock
+    - bt_elev
+    - bt_elev_cancel
+    - color_compressed
+    - move_simple
+    - goal_pose
+    tf_list:
+    - odom_basefootprint
+    srv_server_list:
+    - load_map
     srv_client_list:
-    # - call_elevator
-    # - get_called_elevator
-    # - get_elevator_information
-    # - is_door_opened
-    # - request_door_open
-    # - select_elevator_floor
+    - call_elevator
+    - get_called_elevator
+    - get_elevator_information
+    - is_door_opened
+    - request_door_open
+    - select_elevator_floor
     - add_two_ints
