+++ config/server.yaml
@@ -7,21 +7,21 @@
     rep_port: 25570
     sub_list:
     - cmd_vel
-    # - plan
-    # - footprint
-    # - map
-    # - bt_elev_feedback
-    # - bt_elev_result
-    # - nav2_status
-    # tf_list:
-    # - map_odom
+    - plan
+    - footprint
+    - map
+    - bt_elev_feedback
+    - bt_elev_result
+    - nav2_status
+    tf_list:
+    - map_odom
     srv_server_list:
-    # - call_elevator
-    # - get_called_elevator
-    # - get_elevator_information
-    # - is_door_opened
-    # - request_door_open
-    # - select_elevator_floor
+    - call_elevator
+    - get_called_elevator
+    - get_elevator_information
+    - is_door_opened
+    - request_door_open
+    - select_elevator_floor
     - add_two_ints
-    # srv_client_list:
-    # - load_map
+    srv_client_list:
+    - load_map
