+++ src/bridge_rcl_node.cpp
@@ -13,38 +13,42 @@
  *         SPDX-License-Identifier: MIT
  */
 
-#include "cloud_bridge/bridge_node.hpp"
-#include "cloud_bridge/bridge_client.hpp"
+#include "cloud_bridge/bridge_rcl_node.hpp"
+#include "cloud_bridge/zmq_transport.hpp"
 #include "cloud_bridge/logging.hpp"
 
 #include <rosidl_typesupport_cpp/message_type_support.hpp>
 #include <cstring>
 #include <chrono>
+#include <unistd.h>
 
 #include <std_msgs/msg/string.hpp>
 
-#include <example_interfaces/srv/add_two_ints.hpp>
-
-BridgeNode::BridgeNode(rcl_allocator_t *alloc, rcl_context_t *context, std::string ns)
-    : running(true), alloc(alloc), context(context), node(rcl_get_zero_initialized_node()), wait_set_(rcl_get_zero_initialized_wait_set())
+BridgeRclNode::BridgeRclNode(rcl_allocator_t *alloc, rcl_context_t *context, std::string ns)
+    : running(true), alloc(alloc), context(context), node(rcl_get_zero_initialized_node()), 
+    wait_set_sub_(rcl_get_zero_initialized_wait_set()),
+    wait_set_service_(rcl_get_zero_initialized_wait_set())
 {
+    running = true;
     rcl_node_options_t opts = rcl_node_get_default_options();
     rcl_ret_t rc;
 
-    rc = rcl_node_init(&node, "bridge", ns.c_str(), context, &opts);
+    rc = rcl_node_init(&node, "cloud_bridge_bridge_node", ns.c_str(), context, &opts);
     if (rc != RCL_RET_OK)
     {
         ERROR("rcl_node_init failed: " << rc);
     }
-    thread = std::thread(&BridgeNode::execute, this);
-    auto node_handle = std::make_shared<rclcpp::Node>("tf_bridge", ns.c_str());
+    thread = std::thread(&BridgeRclNode::execute, this);
+    service_thread = std::thread(&BridgeRclNode::execute_service, this);
+    auto node_handle = std::make_shared<rclcpp::Node>("cloud_bridge_tf_node", ns.c_str());
     tf_broadcaster_ = std::make_shared<tf2_ros::TransformBroadcaster>(node_handle);
 }
 
-BridgeNode::~BridgeNode()
+BridgeRclNode::~BridgeRclNode()
 {
     running = false;
     thread.join();
+    service_thread.join();
 
     rcl_ret_t rc;
 
@@ -66,10 +70,16 @@ BridgeNode::~BridgeNode()
         }
     }
 
-    rc = rcl_wait_set_fini(&wait_set_);
+    rc = rcl_wait_set_fini(&wait_set_sub_);
     if (rc != RCL_RET_OK)
     {
-        ERROR("rcl_wait_set_fini failed: " << rc);
+        ERROR("sub rcl_wait_set_fini failed: " << rc);
+    }
+
+    rc = rcl_wait_set_fini(&wait_set_service_);
+    if (rc != RCL_RET_OK)
+    {
+        ERROR("service rcl_wait_set_fini failed: " << rc);
     }
 
     rc = rcl_node_fini(&node);
@@ -79,47 +89,44 @@ BridgeNode::~BridgeNode()
     }
 }
 
-void BridgeNode::execute()
+void BridgeRclNode::execute()
 {
     size_t sub_count = 0;
-    size_t service_count = 0;
     rcl_ret_t rc;
 
     const unsigned int timeout = 100; // msec
+    running = true;
     while (running)
     {
         size_t new_sub_count;
-        size_t new_service_count;
         {
-            std::lock_guard<std::mutex> lock(mutex);
-            while (!actions.empty())
+            std::lock_guard<std::mutex> lock(sub_mutex);
+            while (!add_sub_actions.empty())
             {
-                auto action = actions.front();
+                auto action = add_sub_actions.front();
                 action();
-                actions.pop();
+                add_sub_actions.pop();
             }
             new_sub_count = subscribers.size();
-            new_service_count = service_servers.size();
         }
 
-        if (new_sub_count == 0 && new_service_count == 0)
+        if (new_sub_count == 0)
         {
             std::this_thread::sleep_for(std::chrono::milliseconds(timeout));
             continue;
         }
 
-        if (new_sub_count != sub_count || new_service_count != service_count)
+        if (new_sub_count != sub_count )
         {
             sub_count = new_sub_count;
-            service_count = new_service_count;
 
-            rc = rcl_wait_set_fini(&wait_set_);
+            rc = rcl_wait_set_fini(&wait_set_sub_);
             if (rc != RCL_RET_OK)
             {
                 ERROR("rcl_wait_set_fini failed: " << rc);
             }
 
-            rc = rcl_wait_set_init(&wait_set_, sub_count, 0, 0, 0, service_count, 0, context, *alloc);
+            rc = rcl_wait_set_init(&wait_set_sub_, sub_count, 0, 0, 0, 0, 0, context, *alloc);
             if (rc != RCL_RET_OK)
             {
                 ERROR("rcl_wait_set_init failed: " << rc);
@@ -127,10 +134,10 @@ void BridgeNode::execute()
         }
         else
         {
-            rc = rcl_wait_set_clear(&wait_set_);
+            rc = rcl_wait_set_clear(&wait_set_sub_);
             if (rc != RCL_RET_OK)
             {
-                ERROR("rcl_wait_set_clear failed: " << rc);
+                ERROR("sub rcl_wait_set_clear failed: " << rc);
                 return;
             }
         }
@@ -138,27 +145,16 @@ void BridgeNode::execute()
         for (const auto &it : subscribers)
         {
             size_t index;
-            rc = rcl_wait_set_add_subscription(&wait_set_, &it->sub, &index);
+            rc = rcl_wait_set_add_subscription(&wait_set_sub_, &it->sub, &index);
             if (rc != RCL_RET_OK)
             {
                 ERROR("rcl_wait_set_add_subscription failed: " << rc);
             }
         }
 
-        for (const auto &it : service_servers)
-        {
-            size_t index;
-            rc = rcl_wait_set_add_service(&wait_set_, &it->rcl_service, &index);
-            if (rc != RCL_RET_OK)
-            {
-                ERROR("rcl_wait_set_add_service failed: " << rc);
-            }
-        }
-        
-        rc = rcl_wait(&wait_set_, RCL_MS_TO_NS(timeout));
+        rc = rcl_wait(&wait_set_sub_, RCL_MS_TO_NS(timeout));
         if (rc == RCL_RET_TIMEOUT)
         {
-            // std::cout << "RCL_RET_TIMEOUT: " << sub_count << std::endl;
             continue;
         }
         if (rc != RCL_RET_OK)
@@ -167,120 +163,120 @@ void BridgeNode::execute()
         }
         for (size_t i = 0; i < sub_count; i++)
         {
-            if (wait_set_.subscriptions[i])
+            if (wait_set_sub_.subscriptions[i])
             {
-                Subscriber *sub = (Subscriber *)wait_set_.subscriptions[i];
+                Subscriber *sub = (Subscriber *)wait_set_sub_.subscriptions[i];
                 handle_message(sub);
             }
         }
-        for (size_t i = 0; i < service_count; ++i) {
-            if (wait_set_.services[i])
-            {
-                ServiceServer *ss = (ServiceServer *) wait_set_.services[i];
-                handle_service(ss);
-            }
-        }
-        // std::cout << "-----------------------------" << std::endl;
     }
 }
 
-rmw_qos_profile_t BridgeNode::parseQosString(std::string qos_string)
+void BridgeRclNode::execute_service()
+{
+    size_t service_count = 0;
+    rcl_ret_t rc;
+
+    const unsigned int timeout = 100; // msec
+    while (running)
+    {
+        size_t new_service_count;
+        {
+            std::lock_guard<std::mutex> lock(service_mutex);
+            while (!add_service_actions.empty())
+            {
+                auto action = add_service_actions.front();
+                action();
+                add_service_actions.pop();
+            }
+            new_service_count = service_servers.size();
+        }
+
+        if (new_service_count == 0)
+        {
+            std::this_thread::sleep_for(std::chrono::milliseconds(timeout));
+            continue;
+        }
+
+        if (new_service_count != service_count)
+        {
+            service_count = new_service_count;
+
+            rc = rcl_wait_set_fini(&wait_set_service_);
+            if (rc != RCL_RET_OK)
+            {
+                ERROR("service rcl_wait_set_fini failed: " << rc);
+            }
+
+            rc = rcl_wait_set_init(&wait_set_service_, 0, 0, 0, 0, service_count, 0, context, *alloc);
+            if (rc != RCL_RET_OK)
+            {
+                ERROR("service rcl_wait_set_init failed: " << rc);
+            }
+        }
+        else
+        {
+            rc = rcl_wait_set_clear(&wait_set_service_);
+            if (rc != RCL_RET_OK)
+            {
+                ERROR("service rcl_wait_set_clear failed: " << rc);
+                return;
+            }
+        }
+
+        for (const auto &it : service_servers)
+        {
+            size_t index;
+            rc = rcl_wait_set_add_service(&wait_set_service_, &it->rcl_service, &index);
+            if (rc != RCL_RET_OK)
+            {
+                ERROR("rcl_wait_set_add_service failed: " << rc);
+            }
+        }
+        
+        rc = rcl_wait(&wait_set_service_, RCL_MS_TO_NS(timeout));
+        if (rc == RCL_RET_TIMEOUT)
+        {
+            continue;
+        }
+        if (rc != RCL_RET_OK)
+        {
+            ERROR("rcl_wait failed: " << rc);
+        }
+
+        for (size_t i = 0; i < service_count; ++i) {
+            if (wait_set_service_.services[i])
+            {
+                ServiceServer *service_server = (ServiceServer *) wait_set_service_.services[i];
+                handle_service(service_server);
+            }
+        }
+    }
+}
+
+
+
+rmw_qos_profile_t BridgeRclNode::parseQosString(std::string qos_string)
 {
   rmw_qos_profile_t qos = rmw_qos_profile_default;
   qos.durability = RMW_QOS_POLICY_DURABILITY_TRANSIENT_LOCAL;
   if(qos_string == "sensor_data") {
-    LOG("parseQosString rmw_qos_profile_sensor_data");
     qos = rmw_qos_profile_sensor_data;
   } else if(qos_string == "parameters") {
-    LOG("parseQosString rmw_qos_profile_parameters");
     qos = rmw_qos_profile_parameters;
   } else if(qos_string == "parameter_events") {
-    LOG("parseQosString rmw_qos_profile_parameter_events");
     qos = rmw_qos_profile_parameter_events;
   } else if(qos_string == "system_default") {
-    LOG("parseQosString rmw_qos_profile_system_default");
     qos = rmw_qos_profile_system_default;
   } else if(qos_string == "services_default") {
-    LOG("parseQosString rmw_qos_profile_services_default");
     qos = rmw_qos_profile_services_default;
-  } else{
-    LOG("parseQosString rmw_qos_profile_default");
   }
 
   return qos;
 }
 
-void BridgeNode::remove(BridgeClient *client)
-{
-    auto action = [=]() {
-        for (auto it = subscribers.begin(); it != subscribers.end(); /* empty */)
-        {
-            Subscriber *sub = it->get();
-            if (sub->clients.find(client) != sub->clients.end())
-            {
-                DEBUG("Removing client from subscribers on " << sub->topic << " topic");
-                sub->clients.erase(client);
-                if (sub->clients.empty())
-                {
-                    LOG("Removing subscriber on " << sub->topic << " topic");
-
-                    rcl_ret_t rc = rcl_subscription_fini(&sub->sub, &node);
-                    if (rc != RCL_RET_OK)
-                    {
-                        ERROR("rcl_subscription_fini failed: " << rc);
-                    }
-
-                    it = subscribers.erase(it);
-                }
-                else
-                {
-                    ++it;
-                }
-            }
-            else
-            {
-                ++it;
-            }
-        }
-    };
-
-    {
-        std::lock_guard<std::mutex> lock(mutex);
-        actions.push(action);
-    }
-
-    for (auto it = publishers.begin(), eit = publishers.end(); it != eit; /* empty */)
-    {
-        if (it->second.clients.find(client) != it->second.clients.end())
-        {
-            DEBUG("Removing client from publishers on " << it->first << " topic");
-            it->second.clients.erase(client);
-            if (it->second.clients.empty())
-            {
-                LOG("Removing publisher on " << it->first << " topic");
-
-                rcl_ret_t rc = rcl_publisher_fini(&it->second.pub, &node);
-                if (rc != RCL_RET_OK)
-                {
-                    ERROR("rcl_publisher_fini failed: " << rc);
-                }
-
-                it = publishers.erase(it);
-            }
-            else
-            {
-                ++it;
-            }
-        }
-        else
-        {
-            ++it;
-        }
-    }
-}
-
-void BridgeNode::add_subscriber(const std::string &topic, const std::string &type,
-    BridgeClient *client, std::string &qos)
+void BridgeRclNode::add_subscriber(const std::string &topic, const std::string &type,
+    ZmqTransport *zmq_transport, std::string &qos)
 {
     const MessageType *message_type = types.get(type);
     if (message_type == NULL)
@@ -293,7 +289,6 @@ void BridgeNode::add_subscriber(const std::string &topic, const std::string &typ
             Subscriber *sub = it->get();
             if (sub->topic == topic)
             {
-                sub->clients.insert(client);
                 return;
             }
         }
@@ -302,7 +297,7 @@ void BridgeNode::add_subscriber(const std::string &topic, const std::string &typ
         s->sub = rcl_get_zero_initialized_subscription();
         s->type = message_type;
         s->topic = topic;
-        s->clients.insert(client);
+        s->zmq_transport = zmq_transport;
 
         rcl_subscription_options_t sub_opt = rcl_subscription_get_default_options();
         sub_opt.qos = parseQosString(qos);
@@ -317,20 +312,19 @@ void BridgeNode::add_subscriber(const std::string &topic, const std::string &typ
         auto it = subscribers.insert(subscribers.end(), std::move(s));
         assert((void *)it->get() == (void *)&it->get()->sub);
 
-        LOG("BridgeNode, Subscribed " << type << " on " << topic);
+        LOG("BridgeRclNode, Subscribed " << type << " on " << topic);
     };
 
-    std::lock_guard<std::mutex> lock(mutex);
-    actions.push(action);
+    std::lock_guard<std::mutex> lock(sub_mutex);
+    add_sub_actions.push(action);
 }
 
-void BridgeNode::add_publisher(const std::string &topic, const std::string &type,
-    BridgeClient *client, std::string &qos)
+void BridgeRclNode::add_publisher(const std::string &topic, const std::string &type,
+    ZmqTransport *zmq_transport, std::string &qos)
 {
     auto it = publishers.find(topic);
     if (it != publishers.end())
     {
-        it->second.clients.insert(client);
         return;
     }
 
@@ -354,15 +348,15 @@ void BridgeNode::add_publisher(const std::string &topic, const std::string &type
     Publisher p;
     p.pub = pub;
     p.type = message_type;
-    p.clients.insert(client);
+    p.zmq_transport = zmq_transport;
 
     publishers.insert(std::make_pair(topic, p));
 
     LOG("Publishing " << type << " on " << topic);
 }
 
-void BridgeNode::add_service_server(const std::string& topic, const std::string& type,
-    BridgeClient* client)
+void BridgeRclNode::add_service_server(const std::string& topic, const std::string& type,
+    ZmqTransport* zmq_transport)
 {
     const MessageType *type_req = types.get(type, "request");
     const MessageType *type_res = types.get(type, "response");
@@ -377,7 +371,6 @@ void BridgeNode::add_service_server(const std::string& topic, const std::string&
             ServiceServer *service_server = it->get();
             if (service_server->topic == topic)
             {
-                service_server->clients.insert(client);
                 return;
             }
         }
@@ -387,7 +380,7 @@ void BridgeNode::add_service_server(const std::string& topic, const std::string&
         new_service_server->type_req = type_req;
         new_service_server->type_res = type_res;
         new_service_server->topic = topic;
-        new_service_server->clients.insert(client);
+        new_service_server->zmq_transport = zmq_transport;
 
         rcl_service_options_t service_options = rcl_service_get_default_options();
 
@@ -405,17 +398,16 @@ void BridgeNode::add_service_server(const std::string& topic, const std::string&
         LOG("Service Served " << type << " on " << topic);
     };
 
-    std::lock_guard<std::mutex> lock(mutex);
-    actions.push(action);
+    std::lock_guard<std::mutex> lock(service_mutex);
+    add_service_actions.push(action);
 }
 
-void BridgeNode::add_service_client(const std::string& topic, const std::string& type,
-    BridgeClient* client)
+void BridgeRclNode::add_service_client(const std::string& topic, const std::string& type,
+    ZmqTransport* zmq_transport)
 {
     auto it = service_clients.find(topic);
     if (it != service_clients.end())
     {
-        it->second.clients.insert(client);
         return;
     }
 
@@ -442,35 +434,13 @@ void BridgeNode::add_service_client(const std::string& topic, const std::string&
     service_client.type_req = type_req;
     service_client.type_res = type_res;
     service_client.topic = topic;
-    service_client.clients.insert(client);
+    service_client.zmq_transport = zmq_transport;
     service_clients.insert(std::make_pair(topic, service_client));
 
     LOG("Service Client " << type << " on " << topic);
 }
 
-bool BridgeNode::wait_for_server_to_be_available(
-  rcl_node_t * node, rcl_client_t * client,
-  size_t max_tries, int64_t period_ms)
-{
-    LOG("wait_for_server_to_be_available");
-    size_t iteration = 0;
-    while (iteration < max_tries) {
-        ++iteration;
-        bool is_ready;
-        rcl_ret_t ret = rcl_service_server_is_available(node, client, &is_ready);
-        if (ret != RCL_RET_OK) {
-        ERROR( "Error in rcl_service_server_is_available: %s" << rcl_get_error_string().str);
-        return false;
-        }
-        if (is_ready) {
-        return true;
-        }
-        std::this_thread::sleep_for(std::chrono::milliseconds(period_ms));
-    }
-    return false;
-}
-
-void BridgeNode::publish(const std::string &topic, const std::vector<uint8_t> &data)
+void BridgeRclNode::publish(const std::string &topic, const std::vector<uint8_t> &data)
 {
     if (topic == "tf")
     {
@@ -488,7 +458,7 @@ void BridgeNode::publish(const std::string &topic, const std::vector<uint8_t> &d
             ERROR("failed to serialize serialized message topic tf");
             return;
         }
-        DEBUG("BridgeNode::publish tf_broadcaster_->sendTransform");
+        DEBUG("BridgeRclNode::publish tf_broadcaster_->sendTransform");
         tf_broadcaster_->sendTransform(*transform_msg);
     }
     else
@@ -511,7 +481,7 @@ void BridgeNode::publish(const std::string &topic, const std::vector<uint8_t> &d
                 DEBUG("Unserializing message for " << topic << " topic");
                 if (Unserialize(msg, type->introspection, data))
                 {
-                    DEBUG("BridgeNode rcl_publish topic "<< topic);
+                    DEBUG("BridgeRclNode rcl_publish topic "<< topic);
                     rcl_ret_t rc = rcl_publish(pub, msg, NULL);
                     if (rc != RCL_RET_OK)
                     {
@@ -535,7 +505,7 @@ void BridgeNode::publish(const std::string &topic, const std::vector<uint8_t> &d
     }
 }
 
-void BridgeNode::handle_tf(geometry_msgs::msg::TransformStamped& transform, BridgeClient *client)
+void BridgeRclNode::handle_tf(geometry_msgs::msg::TransformStamped& transform, ZmqTransport *zmq_transport)
 {
     rcl_serialized_message_t serialized_msg_ = rmw_get_zero_initialized_serialized_message();
     size_t tf_msg_size = 0;
@@ -571,11 +541,11 @@ void BridgeNode::handle_tf(geometry_msgs::msg::TransformStamped& transform, Brid
     }
     std::vector<uint8_t> data;
     data.insert(data.end(), serialized_msg_.buffer, serialized_msg_.buffer + serialized_msg_.buffer_length);
-    DEBUG("bridgeNode::handle_message handle_tf publish client->publish");
-    client->publish("tf", "geometry_msgs/TransformStamped", data);
+    DEBUG("BridgeRclNode::handle_tf zmq_transport->send_publish");
+    zmq_transport->send_publish("tf", "geometry_msgs/TransformStamped", data);
 }
 
-void BridgeNode::handle_message(Subscriber *sub)
+void BridgeRclNode::handle_message(Subscriber *sub)
 {
     rcl_ret_t rc;
 
@@ -584,13 +554,10 @@ void BridgeNode::handle_message(Subscriber *sub)
     {
         if (sub->type->init(msg))
         {
-            LOG("msg address 111  " << msg );
             rmw_message_info_t info;
             rc = rcl_take(&sub->sub, msg, &info, NULL);
-            LOG("msg address 222  " << msg );
             if (rc == RCL_RET_OK)
             {
-                // LOG("New message received " << sub->topic << " topic");
                 std::vector<uint8_t> data;
                 // data.reserve(4096);
 
@@ -598,11 +565,8 @@ void BridgeNode::handle_message(Subscriber *sub)
                 Serialize(msg, sub->type->introspection, data);
             
                 std::string target_topic = sub->topic;
-                for (auto client : sub->clients)
-                {
-                    DEBUG("BridgeNode client->publish topic: " << target_topic);
-                    client->publish(target_topic, sub->type->type_string, data);
-                }
+                DEBUG("BridgeRclNode zmq_transport->send_publish topic: " << target_topic);
+                sub->zmq_transport->send_publish(target_topic, sub->type->type_string, data);
             }
             else if (rc == RCL_RET_SUBSCRIPTION_TAKE_FAILED)
             {
@@ -628,7 +592,7 @@ void BridgeNode::handle_message(Subscriber *sub)
     }
 }
 
-void BridgeNode::service(const std::string& topic, 
+void BridgeRclNode::service(const std::string& topic, 
     const std::vector<uint8_t>& req_data, std::vector<uint8_t>& res_data)
 {
     auto it = service_clients.find(topic);
@@ -648,27 +612,26 @@ void BridgeNode::service(const std::string& topic,
     {
         if (type_req->init(req_msg))
         {
-            LOG("Unserializing req_msg for " << topic << " topic");
+            DEBUG("Unserializing req_msg for " << topic << " topic");
             if (Unserialize(req_msg, type_req->introspection, req_data))
             {
-                LOG("BridgeNode rcl_send_request topic: "<< topic);
                 int64_t sequence_number = 0;
                 rcl_ret_t rc = rcl_send_request(rcl_client, req_msg, &sequence_number);
-                LOG("BridgeNode rcl_send_request sequence_number: "<< sequence_number);
+                DEBUG("BridgeRclNode rcl_send_request sequence_number: "<< sequence_number);
                 if (rc != RCL_RET_OK)
                 {
                     ERROR("rcl_send_request failed: " << rc);
                 }
-
-                LOG("BridgeNode wait for rcl_take_response topic "<< topic);
                 rmw_request_id_t request_id;
                 rc = rcl_take_response(rcl_client, &request_id, res_msg);
-                if (rc != RCL_RET_OK)
+                while (rc != RCL_RET_OK)
                 {
-                    ERROR("rcl_take_response failed: " << rc );
+                    std::this_thread::sleep_for(std::chrono::milliseconds(50));
+                    // ERROR("rcl_take_response failed: " << rc );
+                    rc = rcl_take_response(rcl_client, &request_id, res_msg);
                 }
-                LOG("BridgeNode rcl_take_response sequence_number: "<< request_id.sequence_number);
-                LOG("Serialize response msg topic "<< topic);
+                DEBUG("BridgeRclNode rcl_take_response sequence_number: "<< request_id.sequence_number);
+                DEBUG("Serialize response msg topic "<< topic);
                 Serialize(res_msg, type_res->introspection, res_data);
             }
 
@@ -689,46 +652,43 @@ void BridgeNode::service(const std::string& topic,
     }
 }
 
-void BridgeNode::handle_service(ServiceServer* service_server)
+void BridgeRclNode::handle_service(ServiceServer *activate_server)
 {
-    LOG("handle_service " << service_server->topic << " service");
+    DEBUG("handle_service " << activate_server->topic << " service");
     rcl_ret_t rc;
 
-    void *req_msg = malloc(service_server->type_req->size);
-    void *res_msg = malloc(service_server->type_res->size);
+    void *req_msg = malloc(activate_server->type_req->size);
+    void *res_msg = malloc(activate_server->type_res->size);
     if (req_msg)
     {
-        if (service_server->type_req->init(req_msg))
+        if (activate_server->type_req->init(req_msg))
         {
-            LOG("rcl_take_request " << service_server->topic << " service");
+            DEBUG("rcl_take_request " << activate_server->topic << " service");
             rmw_request_id_t request_id; 
-            rc = rcl_take_request(&service_server->rcl_service, &request_id, req_msg);
+            rc = rcl_take_request(&activate_server->rcl_service, &request_id, req_msg);
             if (rc == RCL_RET_OK)
             {
-
                 std::vector<uint8_t> req_data;
                 // req_data.reserve(4096);
 
-                LOG("Serializing req_msg for " << service_server->topic << " service");
-                Serialize(req_msg, service_server->type_req->introspection, req_data);
+                DEBUG("Serializing req_msg for " << activate_server->topic << " service");
+                Serialize(req_msg, activate_server->type_req->introspection, req_data);
 
-                std::string target_topic = service_server->topic;
-                for (auto client : service_server->clients)
+                std::string target_topic = activate_server->topic;
+                auto zmq_transport = activate_server->zmq_transport;
+
+                DEBUG("BridgeRclNode client->send_request_and_get_response topic: " << activate_server->topic);
+                std::vector<uint8_t> res_data;
+                zmq_transport->send_request_and_get_response(target_topic, req_data, res_data);
+
+                DEBUG("Unserializing res_msg for " << activate_server->topic << " topic, res_size: " << res_data.size());
+                if (Unserialize(res_msg, activate_server->type_res->introspection, res_data))
                 {
-                    LOG("BridgeNode client->send_request topic: " << service_server->topic);
-                    std::vector<uint8_t> res_data;
-                    client->send_request(target_topic, req_data, res_data);
-
-                    LOG("Unserializing res_msg for " << service_server->topic << " topic, res_size: " << res_data.size());
-                    LOG("service_server->type_res->size: " <<  service_server->type_res->size);
-                    if (Unserialize(res_msg, service_server->type_res->introspection, res_data))
+                    DEBUG("BridgeRclNode rcl_send_response topic "<< activate_server->topic);
+                    rc = rcl_send_response(&activate_server->rcl_service, &request_id, res_msg);
+                    if (rc != RCL_RET_OK)
                     {
-                        LOG("BridgeNode rcl_send_response topic "<< service_server->topic);
-                        rc = rcl_send_response(&service_server->rcl_service, &request_id, res_msg);
-                        if (rc != RCL_RET_OK)
-                        {
-                            ERROR("rcl_publish failed: " << rc);
-                        }
+                        ERROR("rcl_publish failed: " << rc);
                     }
                 }
             }
@@ -737,12 +697,12 @@ void BridgeNode::handle_service(ServiceServer* service_server)
                 ERROR("rcl_take_request failed: " << rc);
             }
 
-            service_server->type_req->fini(req_msg);
-            service_server->type_res->fini(res_msg);
+            activate_server->type_req->fini(req_msg);
+            activate_server->type_res->fini(res_msg);
         }
         else
         {
-            ERROR("Message init failed on " << service_server->topic << " topic");
+            ERROR("Message init failed on " << activate_server->topic << " topic");
         }
 
         free(req_msg);
@@ -750,6 +710,6 @@ void BridgeNode::handle_service(ServiceServer* service_server)
     }
     else
     {
-        ERROR("Out of memory when receiving message " << service_server->topic << " topic");
+        ERROR("Out of memory when receiving message " << activate_server->topic << " topic");
     }
 }
